+++ lib/td_dd/auth/current_user.ex
@@ -1,13 +1,14 @@
 defmodule TdDd.Auth.CurrentUser do
   @moduledoc false
 
-  alias Guardian.Plug, as: GuardianPlug
-  alias Plug.Conn, as: PlugConn
+  import Guardian.Plug, only: [current_resource: 1]
+
+  alias Plug.Conn
 
   def init(opts), do: opts
 
   def call(conn, _opts) do
-    current_user = GuardianPlug.current_resource(conn)
-    PlugConn.assign(conn, :current_user, current_user)
+    current_user = current_resource(conn)
+    Conn.assign(conn, :current_user, current_user)
   end
 end