+++ test/support/data_structure_case.ex
@@ -21,12 +21,14 @@ defmodule TdDd.DataStructureCase do
       end
 
       def create_hierarchy(names) do
+        %{id: system_id} = insert(:system)
+
         dsvs =
           Enum.map(
             names,
             &insert(:data_structure_version,
               name: &1,
-              data_structure: build(:data_structure, external_id: &1)
+              data_structure: build(:data_structure, external_id: &1, system_id: system_id)
             )
           )